requirements.txt
gradio>=4.44.1
openai==0.28.1
python-dotenv>=1.0.0
